ajout pages notifications

+++ web/notification/suppr_message.php
@@ -7,14 +7,26 @@ require_once(LIB.'/lib_db.class.php');
 
 <?php
 
-$id=$_GET['id'];
-$db = new DB_connection();
+if(isset($_GET['id']))
+{
+	$id=$_GET['id'];
+	$db = new DB_connection();
 
-$req="DELETE from Message where id_message='".$id."'";
+	$req="DELETE from Message where id_message='".$id."'";
 
-$db->DB_query($req);
-	
-$url="index.php";
+	$db->DB_query($req);
+}
+
+$link = $_SERVER["QUERY_STRING"];
+if(!empty($link))
+{
+	$url = "index.php?".$link;
+}
+else
+{	
+	$url = "index.php";
+}
+
 header("Refresh:0;url=$url");
 
 ?>
